Changed the Astroid scene.
Doesnt really look like an astroid scene, but it works none the less. Today was alot of trying to make sure that github was on the right track. I changed the git.ignore files so hopefully we dont run into the same issue. I was trying to look at what it would look like to install on the local computer and I am not sure that I know what to do. A little more difficult than expected.  Other than that I set up github on my home mac and took a look at the requirments document. The requirments document should be comeplete sometime tomorrow.