Found Error for Issues 1-5
I think I found the error. I think the error was that we never switch the row and column in one of the statements to reflect what the board actually say. The statement was GamePiecesArray[game.pieceLastTaken.col, game.pieceLastTaken.row] = selectedGamePiece;   We had put row first in that statement and therefore it would cause the diagnol peices to be weird, while still letting us think that we deleted the right thing. Still needs to be checked though.