Did some work on the Help Menu
Everything is now fuinctioning the way I want it to. Now is just a matter of making it actually look decent and have the material that we will need for it.

Debating actually making a small series of videos for the help menu... But that is something to be considered with the group later today