Worked on help a little bit.
We need to make it so that the help is able to be seen while playing the game and any other time throughout the interface. Should be pretty easy to fix. Also settings screen needs to be able to do the same thing and not overlap the help menu. Or if it does, we need to decide how we want it.